requirements.txt
africastalking==2.0
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.7
cryptography==46.0.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
josepy==2.1.0
kombu==5.5.4
mozilla-django-oidc==4.0.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
responses==0.25.8
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
